Arrange tests according to packages, add copyright header
