+++ modules/core/src/test/java/com/elasticinbox/core/message/id/MessageIdPolicyTest.java
@@ -1,0 +1,79 @@
+/**
+ * Copyright (c) 2011-2012 Optimax Software Ltd.
+ * All rights reserved.
+ * 
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ * 
+ *  * Redistributions of source code must retain the above copyright notice,
+ *    this list of conditions and the following disclaimer.
+ *  * Redistributions in binary form must reproduce the above copyright notice,
+ *    this list of conditions and the following disclaimer in the documentation
+ *    and/or other materials provided with the distribution.
+ *  * Neither the name of Optimax Software, ElasticInbox, nor the names
+ *    of its contributors may be used to endorse or promote products derived
+ *    from this software without specific prior written permission.
+ * 
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+ * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+ * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+ * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE
+ * FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
+ * DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
+ * SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+ * CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
+ * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+package com.elasticinbox.core.message.id;
+
+import static org.junit.Assert.*;
+import static org.hamcrest.Matchers.*;
+
+import java.util.Date;
+import java.util.UUID;
+
+import me.prettyprint.cassandra.utils.TimeUUIDUtils;
+
+import org.junit.Test;
+
+import com.elasticinbox.core.message.id.MessageIdBuilder;
+
+public class MessageIdPolicyTest
+{
+	@Test
+	public void testSentDateMessageIdPolicy()
+	{
+		Date date = new Date();
+		UUID uuid2;
+		long ts2;
+
+		for (int i = 0; i < 10000; i++) {
+			uuid2 = new MessageIdBuilder().setSentDate(date).build();
+			ts2 = TimeUUIDUtils.getTimeFromUUID(uuid2);
+			assertThat(1000L, greaterThan(ts2 - date.getTime()));
+		}
+	}
+	
+	@Test
+	public void testCurrentDateMessageIdPolicy()
+	{
+		UUID uuid;
+		UUID prev = new MessageIdBuilder().build();
+		long ts;
+
+		for (int i = 0; i < 10000; i++) {
+			uuid = new MessageIdBuilder().build();
+			ts = TimeUUIDUtils.getTimeFromUUID(uuid);
+			
+			if(uuid.equals(prev))
+				fail("Not unique, same as previous!");
+
+			prev = uuid;
+			
+			assertThat(100L, greaterThan(System.currentTimeMillis() - ts));
+		}
+	}
+
+}
